+++ src/test/java/io/github/astrapi69/file/zip/Zip4jExtensionsTest.java
@@ -28,13 +28,9 @@ import static org.testng.AssertJUnit.assertEquals;
 import static org.testng.AssertJUnit.assertTrue;
 
 import java.io.File;
+import java.io.FileFilter;
 import java.io.FileNotFoundException;
 import java.io.IOException;
-
-import net.lingala.zip4j.ZipFile;
-import net.lingala.zip4j.exception.ZipException;
-import net.lingala.zip4j.model.enums.CompressionLevel;
-import net.lingala.zip4j.model.enums.CompressionMethod;
 
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.AfterMethod;
@@ -43,6 +39,12 @@ import org.testng.annotations.Test;
 
 import io.github.astrapi69.file.FileTestCase;
 import io.github.astrapi69.file.write.WriteFileExtensions;
+import io.github.astrapi69.io.file.filter.SuffixFileFilter;
+import net.lingala.zip4j.ZipFile;
+import net.lingala.zip4j.exception.ZipException;
+import net.lingala.zip4j.model.ZipParameters;
+import net.lingala.zip4j.model.enums.CompressionLevel;
+import net.lingala.zip4j.model.enums.CompressionMethod;
 
 /**
  * The unit test class for the class {@link Zip4jExtensions}.
@@ -135,6 +137,33 @@ public class Zip4jExtensionsTest extends FileTestCase
 	{
 		final BeanTester beanTester = new BeanTester();
 		beanTester.testBean(Zip4jExtensions.class);
+	}
+
+	@Test
+	public void testZipFilesWithPredicate() throws IOException
+	{
+		String suffix;
+
+		suffix = ".txt";
+		final File testFile1 = new File(this.testDir, "testFindFilesFileString.txt");
+		final File testFile2 = new File(this.testDir, "testFindFilesFileString.tft");
+		final File testFile3 = new File(this.deepDir, "testFindFilesFileString.cvs");
+		WriteFileExtensions.string2File(testFile1, "Its a beautifull day!!!");
+		WriteFileExtensions.string2File(testFile2, "Its a beautifull evening!!!");
+		WriteFileExtensions.string2File(testFile3, "Its a beautifull night!!!");
+		final File zipFile = new File(this.zipDir.getAbsoluteFile(), "Zip4j-predicate.zip");
+
+		final ZipParameters parameters = new ZipParameters();
+		final ZipFile zip4jZipFile = new ZipFile(zipFile);
+		Zip4jExtensions.zipFiles(zip4jZipFile, parameters, this.testDir, (file -> {
+			if (file.isDirectory())
+			{
+				return false;
+			}
+			FileFilter fileFilter = SuffixFileFilter.of(suffix, false);
+			return !fileFilter.accept(file);
+		}));
+		System.out.println(zip4jZipFile);
 	}
 
 	/**